% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mod.R
\name{filterModUI}
\alias{filterModUI}
\alias{filterModServer}
\title{Filters Module}
\usage{
filterModUI(id)

filterModServer(id)
}
\arguments{
\item{id}{An ID string that identifies the namespace of the UI and server components.}
}
\description{
Common filters that will be used across the various tabs
}
