% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_mod.R
\name{tlgModUI}
\alias{tlgModUI}
\alias{tlgModServer}
\title{TLG Module}
\usage{
tlgModUI(id, params = NULL)

tlgModServer(
  id,
  dtlg_fn,
  other_fn,
  n_iter,
  datasets,
  filter_params,
  other_params
)
}
\arguments{
\item{id}{An ID string list identifies the namespace of the UI and server components.}

\item{params}{A named vector of function arguments that are set in the UI to allow
interactivity of creating the table. By default none are selected.}
}
\description{
A module that compares the generation time using thedtlg package with
equivalent functoinality that exists using tidyverse/pharmaverse functionality.

The aim of this module is to show the speed improvement of using
dtlg over the other packages.
}
