% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlg_mod.R
\name{tlgModUI}
\alias{tlgModUI}
\alias{tlgModServer}
\title{TLG Module}
\usage{
tlgModUI(id, params = NULL)

tlgModServer(
  id,
  dtlg_fn,
  other_fn,
  n_iter,
  datasets,
  filter_params,
  other_params
)
}
\arguments{
\item{id}{An ID string list identifies the namespace of the UI and server components.}

\item{params}{A named vector of function arguments that are set in the UI to allow
interactivity of creating the table. By default none are selected.}

\item{dtlg_fn}{Function body of the function to use from the dtlg pcakage}

\item{other_fn}{Function body of the function to use as a comparison output}

\item{n_iter}{(Reactive) Number of iterations to run the comparison over}

\item{datasets}{List of datasets to use as part of the comparison. Must be named
in the same way that they are arguments in `dtlg_fn` and `other_fn`}

\item{filter_params}{A named list of arguments in `dtlg_fn` that can be updated
in the UI for different tables}

\item{other_params}{A named list of additional arguments passed to `dtlg_fn`}
}
\description{
A module that compares the generation time using thedtlg package with
equivalent functoinality that exists using tidyverse/pharmaverse functionality.

The aim of this module is to show the speed improvement of using
dtlg over the other packages.
}
